Offline Support module- Bug fixes
Scenario-1
If the user has one item in the cart and its price is changed later on ,then on visiting the cart again, the items in the cart get cleared.
Issue: The cart icon present on  homepage still shows count of items that were initially present in it.

Scenario-2
If the price of any product is changed and then user clicks on its variant,then he is able to see old price and new price of that product.

Scenario -3

User visits My orders page in offline mode.
Issue: Socket Exception occurs